Added 3x3 as an extension!
Both 2x2 and 3x3 cubes will now be supported with extra options for the 3x3 and other sizes as they get installed! The extra options include Whole Cube Rotations and notations that are not face rotations. Updated the look and feel of the screen including new animations that show you how much it has done, there is no length cap since we have limited the space the moves take while letting it overflow as much as it needs! I hope you enjoy this update, I will be adding commentary come to the next update.