Interactivity handler!
More lines of code than needed but this does the job to handle both 2x2 and 3x3 Rubik's cubes! Animations are added to let users know how the program is processing the scramble, we also have a Stalemate Resolution for situations where both Doubles and Primes are selected for the move instead of just having the Double take priority we randomly decide who takes the move.